chore: clean up repository & use client config to achieve green in CI
