+++ .gitignore
@@ -1,17 +1,9 @@
 build/
-build-wrapper/
+.idea
 .polywrap/
 node_modules/
 wrap
 .DS_Store
-.zos.session
-.openzeppelin/.session
-deployments/
-env/
 dist/
 .env
-bin/
-solc
-coverage/
-coverage.json
-yarn-error.log
+coverage
