+++ packages/safe-contracts-wrapper/polywrap.yaml
@@ -5,10 +5,5 @@ project:
 source:
   module: ./src/index.ts
   schema: ./src/schema.graphql
-  import_abis:
-    - uri: wrap://ens/safe.interface.polywrap.eth
-      abi: ../safe-interface/build/wrap.info
-    - uri: wrap://ens/wraps.eth:ethereum@1.1.0
-      abi: ../../../ethereum/wrapper/build/wrap.info
 extensions:
   build: ./polywrap.build.yaml
