+++ packages/safe-contracts-wrapper/src/__tests__/utils.ts
@@ -1,65 +1,7 @@
-import path from "path";
-import { CoreClientConfig, IWrapPackage } from "@polywrap/client-js";
-import { ensResolverPlugin } from "@polywrap/ens-resolver-plugin-js";
-import {
-  ethereumProviderPlugin,
-  Connection,
-  Connections,
-} from "ethereum-provider-js";
-import { loggerPlugin } from "@polywrap/logger-plugin-js";
-import {
-  ClientConfigBuilder,
-  defaultIpfsProviders,
-} from "@polywrap/client-config-builder-js";
-import { ensAddresses, providers } from "@polywrap/test-env-js";
+import { ClientConfigBuilder } from "@polywrap/client-config-builder-js";
+import { CoreClientConfig } from "@polywrap/core-js";
+import { configure } from "../../client-config"
 
 export function getClientConfig(): CoreClientConfig {
-  const ethereumWrapperPath: string = path.join(
-    path.resolve(__dirname),
-    "..",
-    "..",
-    "..",
-    "..",
-    ".."
-  );
-
-  const ethereumWrapperUri = `wrap://fs/${ethereumWrapperPath}/ethereum/wrapper/build`
-  return new ClientConfigBuilder()
-    .addDefaults()
-    .addEnv("wrap://package/ipfs-resolver", {
-      provider: providers.ipfs,
-      fallbackProviders: defaultIpfsProviders,
-    })
-    .addPackages({
-      "wrap://ens/ens.polywrap.eth": ensResolverPlugin({
-        addresses: { testnet: ensAddresses.ensAddress },
-      }),
-      "wrap://ens/wraps.eth:logger@1.0.0": loggerPlugin({
-        logFunc: (level, message) => {
-          console.log(level, message);
-          return true
-        },
-      }) as IWrapPackage,
-      "wrap://ens/wraps.eth:ethereum-provider@1.1.0": ethereumProviderPlugin({
-        connections: new Connections({
-          networks: {
-            testnet: new Connection({
-              provider: providers.ethereum,
-              signer: "0x90F8bf6A479f320ead074411a4B0e7944Ea8c9C1"
-            }),
-          },
-          defaultNetwork: "testnet",
-        }),
-      }) as IWrapPackage,
-    })
-    .addInterfaceImplementation(
-      "wrap://ens/wraps.eth:ethereum-provider@1.1.0",
-      "wrap://ens/wraps.eth:ethereum-provider@1.1.0"
-    )
-    // @TODO(cbrzn): Remove this once the ENS text record content hash has been updated
-    .addRedirect(
-      "ens/wraps.eth:ethereum@1.1.0",
-      ethereumWrapperUri
-    )
-    .build();
+  return configure(new ClientConfigBuilder()).build();
 }
