+++ packages/safe-contracts-wrapper/workflows/validator.cue
@@ -6,19 +6,19 @@ cases: {
   $2: { data: string },
   $3: { data: string },
   $4: {
-    data: "0x9b1f7F645351AF3631a656421eD2e40f2802E6c0",
+    data: "0x9b1f7f645351af3631a656421ed2e40f2802e6c0",
     error?: _|_,
   }
   $5: {
-    data: "0x67B5656d60a809915323Bf2C40A8bEF15A152e3e",
+    data: "0x67b5656d60a809915323bf2c40a8bef15a152e3e",
     error?: _|_,
   }
   $6: {
-    data: "0x0ddb56f661e5bd05fb252f5bc619f74039cd6d63",
+    data: "0x22c41ef673706fe463259715164121ceb9789228",
     error?: _|_,
   }
   $7: {
-    data: "0xb1eccd188bf7c5708c863b77ed115bc176fb7d92",
+    data: "0xc1ea444c91b9da7cba5693f4794eb832e0ba6f91",
     error?: _|_,
   }
 
@@ -31,14 +31,14 @@ cases: {
   
   estimateGas: {
     $0: {
-      data: "113499",
+      data: "114787",
       error?: _|_,
     }
   }
   
   encode: {
     $0: {
-      data: "0x1688f0b900000000000000000000000067b5656d60a809915323bf2c40a8bef15a152e3e0000000000000000000000000000000000000000000000000000000000000060000000000000000000000000000000000000000000000000000000000000002a0000000000000000000000000000000000000000000000000000000000000000",
+      data: "0x1688f0b90000000000000000000000004bf749ec68270027c5910220ceab30cc284c7ba2000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000030000000000000000000000000000000000000000000000000000000000000000",
       error?: _|_,
     }
   }
