--- packages/safe-factory-wrapper/.gitignore
@@ -1,17 +1,0 @@
-build/
-build-wrapper/
-.polywrap/
-node_modules/
-wrap
-.DS_Store
-.zos.session
-.openzeppelin/.session
-deployments/
-env/
-dist/
-.env
-bin/
-solc
-coverage/
-coverage.json
-yarn-error.log
