+++ packages/safe-factory-wrapper/package.json
@@ -10,9 +10,9 @@
     "build"
   ],
   "scripts": {
-    "build": "polywrap build",
+    "build": "polywrap build -c ./client-config.ts",
     "codegen": "polywrap codegen",
-    "deploy": "polywrap deploy",
+    "deploy": "polywrap deploy -o deployments.json",
     "test": "yarn test:e2e",
     "test:e2e": "yarn test:e2e:codegen && jest src/__tests__/e2e/integration.spec --passWithNoTests --runInBand --verbose",
     "test:e2e:codegen": "polywrap codegen -m ./src/__tests__/types/polywrap.yaml -g ./src/__tests__/types/wrap"
