--- packages/safe-wrapper/.nvmrc
@@ -1,1 +1,0 @@
-v16.7.0
