+++ packages/safe-wrapper/deployments.json
@@ -1,0 +1,13 @@
+[
+  {
+    "name": "ipfs_deploy",
+    "steps": [
+      {
+        "name": "ipfs_deploy",
+        "id": "ipfs_deploy.ipfs_deploy",
+        "input": "wrap://fs/./build",
+        "result": "wrap://ipfs/QmYKLNqqY4sscVXysDz2cNXF1NWF4NNtrRycVUnWWT8KMe"
+      }
+    ]
+  }
+]
